Time is a fractional number actually

+++ service/src/main/kotlin/xyz/lebkuchenfm/api/eventstream/EventDto.kt
@@ -51,7 +51,7 @@ data class PlayerStateDto(
     @Serializable
     data class CurrentlyPlayingDto(
         val song: SongDto,
-        val time: Int,
+        val time: Double,
     )
 
     @Serializable
